--- src/mpoly/ctx_clear.c
@@ -1,17 +1,0 @@
-/*
-    Copyright (C) 2017 Daniel Schultz
-
-    This file is part of FLINT.
-
-    FLINT is free software: you can redistribute it and/or modify it under
-    the terms of the GNU Lesser General Public License (LGPL) as published
-    by the Free Software Foundation; either version 2.1 of the License, or
-    (at your option) any later version.  See <https://www.gnu.org/licenses/>.
-*/
-
-#include "mpoly.h"
-
-void mpoly_ctx_clear(mpoly_ctx_t mctx)
-{
-    return;
-}
